Alterado o verbo HTTP das rotas de inativar para PATCH e ajustado o endereço das rotas DELETE

+++ src/main/java/com/logistica/doisv/services/ProdutoService.java
@@ -196,6 +196,10 @@ public class ProdutoService {
     }
 
     private void excluirImagemProduto(Produto produto) throws GeneralSecurityException, IOException {
+        if (produto.getImagem() == null || produto.getImagem().isBlank()) {
+            return;
+        }
+
         GoogleDriveService.excluirArquivoDrive(produto.getIdProduto().toString(), produto.getClass().getSimpleName());
     }
 }
